pyproject.toml
[project]
name = "bq-guard"
version = "0.1.1"
description = "BigQuery query guard TUI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "textual>=0.55.0",
    "google-cloud-bigquery>=3.14.0",
    "pyyaml>=6.0.1",
    "platformdirs>=4.2.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.4.0"]

[project.scripts]
bq-guard = "bq_guard.app:main"

[tool.pytest.ini_options]
addopts = "-q"
